# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Eventos",
        "description": "API RESTful para gerenciamento de eventos",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Servidor da API"
        }
    ],
    "paths": {
        "/api/eventos": {
            "get": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Lista todos os eventos",
                "description": "Lista todos os eventos",
                "operationId": "8df3c75a189cf773a3eb105ea46e145a",
                "responses": {
                    "200": {
                        "description": "Lista de eventos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Eventos listados com sucesso!"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Evento"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token de autenticação inválido ou não fornecido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Token de autenticação inválido ou não fornecido."
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Cria um novo evento",
                "description": "Cria um novo evento",
                "operationId": "4594ce682158c01a68019d32f26180e1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "descricao",
                                    "data_inicio",
                                    "data_final"
                                ],
                                "properties": {
                                    "descricao": {
                                        "type": "string",
                                        "example": "Workshop de Laravel"
                                    },
                                    "local": {
                                        "type": "string",
                                        "example": "Auditório Principal"
                                    },
                                    "vagas": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "data_inicio": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-12-01 10:00:00"
                                    },
                                    "data_final": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-12-01 18:00:00"
                                    },
                                    "cancelado": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Evento criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Evento criado com sucesso!"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Evento"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Não foi possível criar o evento. Verifique os dados informados."
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token de autenticação inválido ou não fornecido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Token de autenticação inválido ou não fornecido."
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Evento": {
                "required": [
                    "id",
                    "descricao",
                    "data_inicio",
                    "data_final",
                    "cancelado"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "descricao": {
                        "type": "string",
                        "example": "Workshop de Laravel"
                    },
                    "local": {
                        "type": "string",
                        "example": "Auditório Principal",
                        "nullable": true
                    },
                    "vagas": {
                        "type": "integer",
                        "example": 50
                    },
                    "data_inicio": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-12-01 10:00:00"
                    },
                    "data_final": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-12-01 18:00:00"
                    },
                    "cancelado": {
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Token de autenticação Bearer. Use o formato: Bearer <token>. Token de exemplo: 12345"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}